+++ Recipient/blood_request.php
@@ -60,13 +60,23 @@
                   <button type="button" class="close" data-dismiss="modal">&times;</button>
                 </div>
                 <div class="modal-body">
-                  <p>Want to delete?</p>
+                  <p>
+                    <?php if ($row['status'] == 'pending') : ?>
+                      Want to delete this request?
+                    <?php else : ?>
+                      You cannot delete this request as it is already <strong><?php echo $row['status']; ?></strong>.
+                    <?php endif; ?>
+                  </p>
                 </div>
                 <div class="modal-footer">
                   <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
-                  <a href="delete_request.php?id=<?php echo $row['id']; ?>">
-                    <button type="button" class="btn btn-danger">Delete</button>
-                  </a>
+                  <?php if ($row['status'] == 'pending') : ?>
+                    <a href="delete_request.php?id=<?php echo $row['id']; ?>">
+                      <button type="button" class="btn btn-danger">Delete</button>
+                    </a>
+                  <?php else : ?>
+                    <button type="button" class="btn btn-danger" disabled>Cannot Delete</button>
+                  <?php endif; ?>
                 </div>
               </div>
             </div>
